use integer type u32 instead of usize

+++ src/solution/s001.rs
@@ -3,8 +3,8 @@ extern crate threadpool;
 use std::sync::mpsc::channel;
 use threadpool::ThreadPool;
 
-pub fn run_concurrently(n1: usize, n2: usize, max_num: usize, num_of_workers: usize) -> usize {
-    let pool = ThreadPool::new(num_of_workers);
+pub fn run_concurrently(n1: u32, n2: u32, max_num: u32, num_of_workers: u32) -> u32 {
+    let pool = ThreadPool::new(num_of_workers as usize);
     let (sender, receiver) = channel();
     for n in 0..max_num {
         let sender = sender.clone();
@@ -19,11 +19,11 @@ pub fn run_concurrently(n1: usize, n2: usize, max_num: usize, num_of_workers: us
                 .expect("channel will be there waiting for the pool");
         });
     }
-    receiver.iter().take(max_num).sum()
+    receiver.iter().take(max_num as usize).sum()
 }
 
-pub fn run(n1: usize, n2: usize, max_num: usize) -> usize {
-    let mut sum: usize = 0;
+pub fn run(n1: u32, n2: u32, max_num: u32) -> u32 {
+    let mut sum: u32 = 0;
     for n in 0..max_num {
         if (n >= n1 && n % n1 == 0) || (n >= n2 && n % n2 == 0) {
             sum += n;
